# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/item-pickup": {
            "get": {
                "tags": [
                    "All item-pickup"
                ],
                "summary": "All item-pickup",
                "description": "Returns all item-pickup",
                "operationId": "retrieve-item-pickup",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "request item-pickup"
                ],
                "summary": "request item-pickup",
                "description": "request item-pickup",
                "operationId": "request-item-pickup",
                "requestBody": {
                    "content": {
                        "application/form-data": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "phone_number",
                                    "email",
                                    "collection_address",
                                    "postal_code",
                                    "package_type",
                                    "commodity_type",
                                    "collection_date",
                                    "collection_time",
                                    "item_description",
                                    "estimated_weight",
                                    "number_of_boxes",
                                    "receiver_first_name",
                                    "receiver_last_name",
                                    "receiver_address",
                                    "receiver_city",
                                    "receiver_state",
                                    "receiver_phone_number",
                                    "receiver_phone_number2",
                                    "delivery_type",
                                    "payment_type"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "text"
                                    },
                                    "last_name": {
                                        "type": "test"
                                    },
                                    "phone_number": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "text"
                                    },
                                    "collection_address": {
                                        "type": "text"
                                    },
                                    "postal_code": {
                                        "type": "text"
                                    },
                                    "package_type": {
                                        "type": "text"
                                    },
                                    "commodity_type": {
                                        "type": "text"
                                    },
                                    "collection_date": {
                                        "type": "text"
                                    },
                                    "collection_time": {
                                        "type": "text"
                                    },
                                    "item_description": {
                                        "type": "text"
                                    },
                                    "estimated_weight": {
                                        "type": "text"
                                    },
                                    "number_of_boxes": {
                                        "type": "text"
                                    },
                                    "receiver_first_name": {
                                        "type": "text"
                                    },
                                    "receiver_last_name": {
                                        "type": "text"
                                    },
                                    "receiver_address": {
                                        "type": "text"
                                    },
                                    "receiver_city": {
                                        "type": "text"
                                    },
                                    "receiver_state": {
                                        "type": "text"
                                    },
                                    "receiver_phone_number": {
                                        "type": "text"
                                    },
                                    "receiver_phone_number2": {
                                        "type": "text"
                                    },
                                    "delivery_type": {
                                        "type": "text"
                                    },
                                    "payment_type": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "request create Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "request create Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Not processed",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/item-pickup/{tracking_id}": {
            "get": {
                "tags": [
                    "Track package"
                ],
                "summary": "",
                "description": "Track Package",
                "operationId": "item-pickup-details",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tracking ID of item-pickup",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Not processed",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}